requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
pickle-mixin